TChem: Moving the RHS and Jacobian functions to the core.
